refactor: 영화 상세 정보 모달의 컴포넌트들을 폴더 분리
