+++ src/components/movieDetails/MoviePoster.ts
